+++ npbench/benchmarks/cavity_flow/cavity_flow_dpnp.py
@@ -31,7 +31,7 @@ def pressure_poisson(nit, p, dx, dy, b):
 def cavity_flow(nx, ny, nt, nit, u, v, dt, dx, dy, p, rho, nu):
     un = np.empty_like(u)
     vn = np.empty_like(v)
-    b = np.zeros((ny, nx))
+    b = np.zeros((ny, nx), dtype=np.float32)
 
     for n in range(nt):
         un[:] = u.copy()
